+++ src/common.h
@@ -5,9 +5,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#define CORNAMUSA_VERSION "0.11.2"
+#define CORNAMUSA_VERSION "0.11.3"
 #define CORNAMUSA_VERSION_MAJOR 0
 #define CORNAMUSA_VERSION_MINOR 11
-#define CORNAMUSA_VERSION_PATCH 2
+#define CORNAMUSA_VERSION_PATCH 3
 
 #endif /* CORNAMUSA_COMMON_H */